fix: Fix joining and leaving conversations in route changes
When the route changed the handler in the left sidebar left the previous
conversation and joined the new one, if any. However, if there was no
sidebar (for example, for guest users) the conversations were not left
or joined on route changes, even if the UI was updated (which led to
issues like calls still being active on the background). Now leaving and
joining conversations on route changes was moved to the main App so it
no longer depends on having a left sidebar. Clearing the token when the
conversation was left was also moved to the App for consistency.

In the CallPhoneDialog component it was explicitly waited for the
conversation to be joined to then join the call, but that would have
required waiting for the conversation to be joined on the route change,
which differed from the existing behaviours in other places. Instead now
a listener for the event sent when the conversation is joined is
listened to to then join the call.

Note that none of this affects Talk when embedded in the Files app or
share pages, as in those cases no router is used and a single
conversation is used, which is explicitly joined when Talk is loaded.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>