refactor: Extract unrelated statements from try/catch
"startPhoneCall" should not throw any exception, as it is internally
guarded by a try/catch. Moreover, even if it threw an exception it would
not be caught by the outer catch anyway, as it is an async function but
it is called without await (because closing the modal should not be
delayed until the call is joined). Therefore it is extracted from the
try/catch to clarify that, in case of an error joining the call, the
conversation will not be removed.

Besides that, in the unlikely case that closing the modal caused an
exception to be thrown that should not in turn cause the conversation to
be removed either.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>