feat: Show live transcriptions in web UI
The transcripts of each participant is shown in its own block with the
avatar and name of the participant, and whenever a transcript for a
different participant arrives a new block is shown.

The transcript area shows four lines of text, which may include the
participant name; the name will be hidden once four or more lines of
text for the same participant are added. When a new line is added the
whole text is immediately scrolled to show the new line.

Using a separate span for each transcript chunk is not strictly needed,
but it will be used in following commits.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>