feat: Set separator between transcript chunks from the language metadata
Different languages use different separators (for example, Chinese does
not use any between kanjis). This detail is included in the language
metadata provided by the "live_transcription" app, so the separator
added between each transcript chunk now respects the language; in case
of a language switch a space is always added.

The language metadata is explicitly loaded if it was not available yet
before the live transcription is enabled to ensure that it will be
available when the transcript is shown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>