feat: Set text direction for the transcript from the language metadata
The text direction of a language is included in the metadata provided by
the "live_transcription" app, so now the transcripts are shown in the
right direction (the characters themselves were already shown in the
right direction due to the Unicode bidi support, but they could be
aligned to the wrong side depending on the main text direction of the
UI).

Note that the whole transcript block is affected, so also the name and
avatar of the author will be affected by the text direction. Due to that
a new block is now added when the text direction changes.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>