fix: Ensure that lines will not disappear when updating last chunk
Now that the last chunk can be replaced with a different one it could
happen that the new one is shorter than the previous one and the total
number of lines is reduced. To prevent all the lines from moving down
if the last one(s) are removed the "min-height" property now guarantees
that the transcript block will not reduce its height even if the number
of lines is reduced by forcing it to the largest known height of the
element.

Nevertheless, the "min-height" property is reset when the whole
transcript is resized, as in that case the text is expected to be
adjusted to the new size and therefore the number of lines is expected
to change.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>