fix: Send initial media state in the LocalStateBroadcaster
Besides changing where they are sent in the code the new code also
reduces the number of signaling messages, as now it sends them directly
to each participant rather than to every participant. Moreover, when
Janus is used, the state is sent now when the remote participant joins
the call rather than when a connection was established with that
participant (or joined, if that participant was not a publisher, as in
that case no connection was established with it). Sending the state when
a connection was established with the remote participant did not
guarantee in any way that the signaling message would be received by the
remote participant (hence the repeated sending with an exponential
backoff), so sending it instead when the participant joins has
essentially the same result (as the repeated sending with an exponential
backoff is kept), but simplifies the code and makes more sense
conceptually (as before the state of the publisher was sent when a
connection with a subscriber was established). Similarly, when Janus is
not used the state is sent now the first time that a connection is
established with the remote participant, but not on every change to the
connected state, as doing it only the first time is actually needed.

Data channel messages are also sent now directly to the other
participant when Janus is not used. However, when Janus is used they are
still sent to all participants, as Janus does not support sending data
channel messages to a specific participant. Unfortunately that means
that data channel messages are still repeated for every participant in
the call when another remote participant joins it. This is specially bad
when the local participant joins a call, as in that case all the remote
participants joined at the same time from the point of view of the local
participant.

Independently of that, note that when Janus is used the media state is
always sent even if the local participant is not publishing, but if
Janus is not used it will not be sent if neither the local participant
nor the remote participant are publishing. While it should be possible
to not sent the media state in all cases if the local participant is not
publishing (because the audio and video will be implicitly disabled in
that case) this is not properly handled by the iOS client yet, so for
now the media state is still sent in that case.

Note that this means that without Janus and when neither the local nor
remote participants are publishing the iOS client will show the audio as
available. In the previous code this was less problematic, because as
the media state was sent to all participants, the media state could be
received by a remote participant (through a signaling message) even if
neither the remote participant nor the local participant was publishing
if the local participant established a connection with a different
remote participant that was publishing. This is no longer the case, but
is nevertheless something to be fixed in the iOS client (and not a big
deal anyway, as limited publishing permissions are unlikely to be used
in a setup without Janus).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>