fix: Send initial name using LocalStateBroadcaster
Similarly to the initial sending of the media state the nick is now sent
using the LocalStateBroadcaster subclasses as part of the current state.
Like for the media state the new code reduces the messages sent by
sending them directly to each participant instead of to all
participants, except for data channels using Janus which are still sent
to all participants.

When Janus is not used the nick is sent only if no connection will be
established with the remote participant, as otherwise the nick is sent
as part of the offer/answer. Nevertheless, the detection of the peer
being set to null does not properly work due to the peer attribute being
initialized to "null", but that will be fixed in a follow up commit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>