fix: Always use high temporal layer in simulcast
When using simulcast the clients may also enable the temporal layer of
the VP8 codec, so when getting the same spatial and temporal layer the
video may not only reduce its resolution, but also its frame rate. This
can look bad specially when there are just a few participants in the
call, so for now (although this might be adjusted in the future for some
scenarios) the high temporal layer is always used independently of the
current spatial layer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>